More work on XSSF Hyperlinks. Still not quite there, but mostly now. Just need to fix a openxml4j bug first though
git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@645179 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFHyperlink.java
@@ -130,7 +130,7 @@ public class HSSFHyperlink implements Hyperlink {
      *
      * @return the 0-based column of the first cell that contains the hyperlink
      */
-    public short getFirstColumn(){
+    public int getFirstColumn(){
         return record.getFirstColumn();
     }
 
@@ -139,8 +139,8 @@ public class HSSFHyperlink implements Hyperlink {
      *
      * @param col the 0-based column of the first cell that contains the hyperlink
      */
-    public void setFirstColumn(short col){
-        record.setFirstColumn(col);
+    public void setFirstColumn(int col){
+        record.setFirstColumn((short)col);
     }
 
     /**
@@ -148,7 +148,7 @@ public class HSSFHyperlink implements Hyperlink {
      *
      * @return the 0-based column of the last cell that contains the hyperlink
      */
-    public short getLastColumn(){
+    public int getLastColumn(){
         return record.getLastColumn();
     }
 
@@ -157,8 +157,8 @@ public class HSSFHyperlink implements Hyperlink {
      *
      * @param col the 0-based column of the last cell that contains the hyperlink
      */
-    public void setLastColumn(short col){
-        record.setLastColumn(col);
+    public void setLastColumn(int col){
+        record.setLastColumn((short)col);
     }
 
     /**
